requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
torch==2.1.0
transformers==4.44.0
qwen-vl-utils==0.0.6
opencv-python-headless==4.9.0
numpy==1.24.3
pillow==10.0.0
pydantic==2.5.0